move class to __init__.py
