+++ extras/devices/__init__.py
@@ -1,5 +1,5 @@
 from devices.base import AbstractDevice, logger
-from devices.LVFS.LVFS import LVFSDevice, LenovoNB
+from devices.LVFS import LVFSDevice, LenovoNB
 from devices.OEM.OEM import OEMDevice, HPEDevice
 
 __all__ = [
